requirements.txt
git+git://github.com/isu-avista/data.git
wheel==0.32.3
pyyaml==5.3.1
python-dotenv==0.15.0
statsmodels==0.12.1
numpy==1.19.4
setuptools>=50.3.2
Werkzeug~=1.0.1
sphinx>=3.3.1
rinohtype>=0.4.2
SQLAlchemy==1.3.20
coverage==5.3
smbus2==0.4.1
PyMLX90614==0.0.3
mpu6050-raspberrypi==1.1
scipy==1.5.4
adafruit-circuitpython-dht==3.5.4
adafruit-circuitpython-mcp3xxx==1.4.3
RPi.GPIO==0.7.0
requests~=2.25.0
pytest>=6.2.2
codecov>=2.1.11
pytest-cov>=2.11.1
fake-rpi>=0.7.1
smbus2==0.4.1
smbus==1.1.post2
psycopg2==2.8.6
